+++ src/dublib/web_requestor/config/constants.py
@@ -1,0 +1,4 @@
+from typing import Literal
+
+AUTHORIZATION_HEADER: Literal["authorization"] = "authorization"
+USER_AGENT_HEADER: Literal["user-agent"] = "user-agent"